split and move posts component
