--- src/components/blog/posts.js
@@ -1,161 +1,0 @@
-import React, { Component } from 'react';
-import { connect } from 'react-redux';
-import { Link } from 'react-router-dom';
-import * as ACTIONS from '../../store/actions/actions';
-import Pagination from 'react-js-pagination';
-import moment from 'moment';
-import withErrorHandler from '../../hoc/withErrorHandler';
-import axios from '../../axios';
-import {
-    Card,
-    CardContent,
-    CardHeader,
-    Button,
-    TextField
-} from '@material-ui/core';
-
-const RenderPosts = ({post}) => {
-    return (
-        <div className="CardStyles">
-            <Card style={{width:'500px', height: '200px', marginBottom: '10px', paddingBottom: '80px'}}>
-                <CardHeader
-                title={<Link to={{ pathname: '/post/' + post.pid, state: {post} }} >
-                        {post.title}
-                        </Link>}
-                subheader={
-                    <div className='FlexColumn'>
-                        <div className='FlexRow'>
-                            {moment(post.date_created).format('MMMM Do, YYYY | h:mm:ss a')}
-                        </div>
-                        <div className='FlexRow'>
-                            <Link style={{paddingLeft: '5ps', textDecoration: 'none'}}
-                                to={{pathname: '/user/' + post.author, state:{post}}}>
-                                By: {post.author}
-                            </Link>
-                        </div>
-                        <div className="FlexRow">
-                            <i className="material-icons">thumb_up</i>
-                            <div className="notification-num-posts">
-                                {post.likes}
-                            </div>
-                        </div>
-                    </div>
-                }
-                />
-                <br />
-                <CardContent>
-                <span style={{overflow: 'hidden'}}>{post.body}</span>
-                </CardContent>
-            </Card>
-        </div>
-    );
-}
-
-class Posts extends Component {
-    constructor(props) {
-        super(props);
-        this.state = {
-            opacity: 0,
-            activePage: 1,
-            PostsPerPage: 5,
-            postsSlice: [],
-        }
-        this.slicePosts = this.slicePosts.bind(this);
-        this.handlePageChange = this.handlePageChange.bind(this);
-        this.handleSearch = this.handleSearch.bind(this);
-    }
-
-    componentDidMount() {
-        this.handleTransition();
-        this.props.onFetchPosts();
-    }
-    handleTransition () {
-        setTimeout(() => this.setState({opacity: 1}), 400);
-    }
-    addPostsToState() {
-        this.animatePosts();
-    }
-    handleSearch(event) {
-        const query = event.target.value;
-        this.props.onFetchPosts({query});
-    }
-
-    slicePosts() {
-        const indexOfLastPost = this.state.activePage * this.state.PostsPerPage;
-        const indexOfFirstPost = indexOfLastPost - this.state.PostsPerPage;
-        this.setState({
-            postsSlice: this.props.posts.slice(indexOfFirstPost, indexOfLastPost)
-        });
-    }
-
-    handlePageChange(pageNumber) {
-        this.setState({activePage: pageNumber});
-        setTimeout(() => this.slicePosts(), 50);
-        setTimeout(() => this.animatePosts(), 100);
-    }
-    render() {
-        return (
-            <div>
-                <div style={{opacity: this.state.opacity, transition: 'opacity 2s ease'}}>
-                    <br />
-
-                {this.props.isAuthenticated ?
-                    <Link to='/addpost' >
-                        <Button variant='contained' color='primary'>
-                            Add Post
-                        </Button>
-                    </Link>
-                    : <Link to='/signup' >
-                        <Button variant='contained' color='primary'>
-                            Signup to Add Post
-                        </Button>
-                    </Link>
-                    
-                }
-                </div>
-                <div>
-                    <TextField
-                        id='search'
-                        label='Search'
-                        margin='normal'
-                        onChange={this.handleSearch}
-                    />
-                </div>
-                <div style={{opacity: this.state.opacity, transition: 'opacity 2s ease'}}>
-                    <h1>Posts</h1>
-                    <div>
-                    {   this.props.posts
-                        ? this.props.posts.map(
-                                post => <RenderPosts opacity={this.state.opacity} key={post.pid} post={post} />
-                        )
-                        : null
-                    }
-                    </div>
-                    <Pagination
-                        activePage={this.state.activePage}
-                        itemsCountPerPage={5}
-                        totalItemsCount={this.props.posts.length}
-                        pageRangeDisplayed={this.props.posts.length/5}
-                        onChange={this.handlePageChange}
-                    />
-                </div>
-            </div>
-        );
-    }
-}
-
-const mapStateToProps = state => {
-    return {
-        posts: state.post.posts,
-        searchPosts: state.post.searchPosts,
-        isAuthenticated: state.auth.isAuthenticated
-    };
-}
-
-const mapDispatchToProps = dispatch => {
-    return {
-        onFetchPosts: query => dispatch(ACTIONS.fetchPosts(query))
-    };
-}
-
-export default connect(mapStateToProps, mapDispatchToProps)(withErrorHandler(Posts, axios));
